pyproject.toml
[tool.poetry]
    name         = "mypypi2"
    version      = "0.1.0"
    description  = "Simple pull-through PyPi cache"
    license      = "MIT"
    readme       = "README.md"
    authors      = ["Nathan Vaughn <nath@nvaughn.email>"]
    homepage     = "https://github.com/NathanVaughn/mypypi2"
    repository   = "https://github.com/NathanVaughn/mypypi2.git"
    package-mode = false

[tool.poetry.dependencies]
    python           = "^3.12"
    flask            = "^3.0.3"    # web framework
    flask-sqlalchemy = "^3.1.1"    # ORM
    pyjson5          = "^1.6.6"    # fast json parser
    requests         = "^2.32.3"   # HTTP requests
    loguru           = "^0.7.2"    # logging
    s3fs             = "^2024.6.1" # S3 filesystem
    lxml             = "^5.3.0"    # fast XML parser
    packaging        = "^24.1"     # PEP 440 version parsing

[tool.poetry.group.dev.dependencies]
    pytest                  = "8.3.2"
    pytest-cov              = "5.0.0"
    pre-commit              = "3.8.0"
    pytest-flask-sqlalchemy = "^1.1.0"
    freezegun               = "^1.5.1" # time mocking

[tool.coverage.run]
    omit = ["tests/*"]

[tool.pyright]
    typeCheckingMode           = "basic"
    venvPath                   = "."
    venv                       = ".venv"
    reportMissingParameterType = true

[tool.ruff]
    # i have an ultra-wide monitor and makes things easier to read
    line-length = 256

[build-system]
    requires      = ["poetry-core"]
    build-backend = "poetry.core.masonry.api"
